refactor: 포멧을 변환하는 함수 한 파일에 정리하기 위해 파일명 변경
convertFormat으로 변경한 후, DateFormat, TimeFormat 위치 이동
